Added logo, added support for importing images, and changed the name of customTypes to custom-types
