+++ @types/index.d.ts
@@ -1,4 +1,4 @@
-declare module 'customTypes' {
+declare module 'custom-types' {
   export interface Devfile {
     name: string;
     version: string;
@@ -25,4 +25,26 @@ declare module 'customTypes' {
     state: boolean,
     freq: number
   }
+}
+
+declare module '*.svg' {
+  import React = require('react')
+  export const ReactComponent: React.SFC<React.SVGProps<SVGSVGElement>>
+  const src: string
+  export default src
+}
+
+declare module '*.jpg' {
+  const content: string
+  export default content
+}
+
+declare module '*.png' {
+  const content: string
+  export default content
+}
+
+declare module '*.json' {
+  const content: string
+  export default content
 }
